\name{EnsDbLite.Hsapiens.81}
\docType{package}

\alias{EnsDbLite.Hsapiens.81-package}
\alias{EnsDbLite.Hsapiens.81}
\alias{EnsDbLite.Hsapiens.81}


\title{Ensembl-based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Sat Oct 31 10:21:24 2015 and based on the GRCh38
}

\author{Nobody}



\examples{
## load the library
##library(EnsDbLite.Hsapiens.81)
## list the contents that are loaded into memory
ls('package:EnsDbLite.Hsapiens.81')
## show the db object that is loaded by calling it's name
EnsDbLite.Hsapiens.81

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
